Use a try-with-resource construction
The code did not close the reader after the file was read. The most
elegant way to do this in modern Java is with the construction in this
commit. It will ensure that the reader is closed even when an exception
occurs (coulc also be done in a finally block).